Move util testing package into e2e/util
