+++ test/e2e/BUILD.bazel
@@ -54,6 +54,7 @@ filegroup(
         ":package-srcs",
         "//test/e2e/framework:all-srcs",
         "//test/e2e/suite:all-srcs",
+        "//test/e2e/util:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
