+++ test/e2e/suite/issuers/acme/certificate/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//test/e2e/framework/addon/pebble:go_default_library",
         "//test/e2e/framework/addon/tiller:go_default_library",
         "//test/e2e/suite/issuers/acme/dnsproviders:go_default_library",
-        "//test/util:go_default_library",
+        "//test/e2e/util:go_default_library",
         "//test/util/generate:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
