+++ test/e2e/suite/issuers/ca/certificate.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
 	"github.com/jetstack/cert-manager/test/e2e/framework"
-	"github.com/jetstack/cert-manager/test/util"
+	"github.com/jetstack/cert-manager/test/e2e/util"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
