+++ test/e2e/suite/issuers/selfsigned/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     deps = [
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
         "//test/e2e/framework:go_default_library",
-        "//test/util:go_default_library",
+        "//test/e2e/util:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
