+++ test/e2e/suite/issuers/vault/BUILD.bazel
@@ -15,7 +15,7 @@ go_library(
         "//test/e2e/framework/addon/tiller:go_default_library",
         "//test/e2e/framework/addon/vault:go_default_library",
         "//test/e2e/suite/issuers/vault/certificate:go_default_library",
-        "//test/util:go_default_library",
+        "//test/e2e/util:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
     ],
