+++ test/util/BUILD.bazel
@@ -1,29 +1,3 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["util.go"],
-    importpath = "github.com/jetstack/cert-manager/test/util",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/apis/certmanager/v1alpha1:go_default_library",
-        "//pkg/client/clientset/versioned/scheme:go_default_library",
-        "//pkg/client/clientset/versioned/typed/certmanager/v1alpha1:go_default_library",
-        "//pkg/util:go_default_library",
-        "//pkg/util/pki:go_default_library",
-        "//vendor/github.com/golang/glog:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/api/extensions/v1beta1:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1beta1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes/typed/core/v1:go_default_library",
-    ],
-)
-
 filegroup(
     name = "package-srcs",
     srcs = glob(["**"]),
